Develop

+++ src/main/java/com/example/project/studentsystem/service/ProfessionService.java
@@ -5,11 +5,9 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.example.project.studentsystem.IService.IProfessionService;
 import com.example.project.studentsystem.dto.CounselorProfessionRelResp;
 import com.example.project.studentsystem.dto.ProfessionResp;
+import com.example.project.studentsystem.entry.*;
 import com.example.project.studentsystem.entry.Class;
-import com.example.project.studentsystem.entry.CounselorProfessionRel;
-import com.example.project.studentsystem.entry.Profession;
-import com.example.project.studentsystem.mapper.ClassMapper;
-import com.example.project.studentsystem.mapper.ProfessionMapper;
+import com.example.project.studentsystem.mapper.*;
 import com.google.common.collect.Lists;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,6 +27,15 @@ public class ProfessionService {
     @Autowired
     private ClassMapper classMapper;
 
+    @Autowired
+    private CounselorProfessionRelMapper counselorProfessionRelMapper;
+
+    @Autowired
+    private StudentMapper studentMapper;
+
+    @Autowired
+    private CourseMapper courseMapper;
+
 
     /**
      * 添加专业
@@ -37,7 +44,7 @@ public class ProfessionService {
      */
     public int addProfession(Profession profession){
 
-        //判断有无该专业，五则添加
+        //判断有无该专业，无则添加
         QueryWrapper<Profession> wrapper = new QueryWrapper<>();
         wrapper.eq("profession_name",profession.getProfessionName());
         List<Profession> professions = professionMapper.selectList(wrapper);
@@ -93,8 +100,67 @@ public class ProfessionService {
      * @return
      */
     public int deleteById(String id){
-      return professionMapper.deleteById(Long.valueOf(id));
+
+        //判断辅导员-专业关联表
+        QueryWrapper<CounselorProfessionRel> counselorProfessionRelQueryWrapper = new QueryWrapper<>();
+        counselorProfessionRelQueryWrapper.eq("profession_id",Long.valueOf(id));
+        List<CounselorProfessionRel> counselorProfessionRels = counselorProfessionRelMapper.selectList(counselorProfessionRelQueryWrapper);
+        if(CollectionUtil.isNotEmpty(counselorProfessionRels)){
+            return -1;
+        }
+
+        //判断学生-专业关联表
+        QueryWrapper<Student> studentQueryWrapper = new QueryWrapper<>();
+        studentQueryWrapper.eq("profession_id",Long.valueOf(id));
+        List<Student> students = studentMapper.selectList(studentQueryWrapper);
+        if(CollectionUtil.isNotEmpty(students)){
+            return -1;
+        }
+        //判断课程表
+        QueryWrapper<Course> courseQueryWrapper = new QueryWrapper<>();
+        courseQueryWrapper.eq("profession_id",Long.valueOf(id));
+        List<Course> courseList = courseMapper.selectList(courseQueryWrapper);
+        if(CollectionUtil.isNotEmpty(courseList)){
+            return -1;
+        }
+
+        QueryWrapper<Class> classWrapper = new QueryWrapper<>();
+        classWrapper.eq("profession_id",Long.valueOf(id));
+
+        classMapper.delete(classWrapper);
+
+        return professionMapper.deleteById(Long.valueOf(id));
     }
 
+
+    /**
+     * 根据条件查询
+     * @param resp
+     * @return
+     */
+    public List<ProfessionResp> findByCondition(ProfessionResp resp){
+        List<ProfessionResp> resultList = Lists.newArrayList();
+        QueryWrapper<Profession> professionWrapper = new QueryWrapper<>();
+        if(resp.getProfessionName()!=null && resp.getDepartment()==null){
+            professionWrapper.eq("profession_name",resp.getProfessionName());
+        }else if(resp.getProfessionName()==null && resp.getDepartment()!=null){
+            professionWrapper.eq("department",resp.getDepartment());
+        }else if(resp.getProfessionName()!=null && resp.getDepartment()!=null){
+            professionWrapper.eq("profession_name",resp.getProfessionName());
+            professionWrapper.eq("department",resp.getDepartment());
+        }
+
+        List<Profession> professions = professionMapper.selectList(professionWrapper);
+        if(CollectionUtil.isNotEmpty(professions)){
+            professions.forEach(profession -> {
+                ProfessionResp resp1 = new ProfessionResp();
+                BeanUtils.copyProperties(profession,resp1);
+                resp1.setId(profession.getId().toString());
+                resultList.add(resp1);
+            });
+        }
+
+        return resultList;
+    }
 
 }
